# openapi.json
{
  "components": {
    "schemas": {
...
      "BookGreeksRequest": {
        "properties": {
          "betas": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Betas",
            "type": "object"
          },
          "book_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Ref"
          },
          "positions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PositionIn"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Positions"
          }
        },
        "title": "BookGreeksRequest",
        "type": "object"
      },
      "BookGreeksResponse": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "risk_free_rate_note": {
            "title": "Risk Free Rate Note",
            "type": "string"
          },
          "stress_grid": {
            "$ref": "#/components/schemas/StressGridOut"
          },
          "underlyings": {
            "items": {
              "$ref": "#/components/schemas/UnderlyingGreeksOut"
            },
            "title": "Underlyings",
            "type": "array"
          }
        },
        "required": [
          "underlyings",
          "stress_grid",
          "risk_free_rate_note",
          "as_of"
        ],
        "title": "BookGreeksResponse",
        "type": "object"
      },
...
      "ChainResponse": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "missing": {
            "title": "Missing",
            "type": "boolean"
          },
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/OptionQuoteOut"
            },
            "title": "Quotes",
            "type": "array"
          },
          "smile": {
            "items": {
              "$ref": "#/components/schemas/ExpirySmile"
            },
            "title": "Smile",
            "type": "array"
          },
          "spot": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "underlier": {
            "title": "Underlier",
            "type": "string"
          }
        },
        "required": [
          "underlier",
          "as_of",
          "spot",
          "stale",
          "quotes",
          "smile",
          "missing"
        ],
        "title": "ChainResponse",
        "type": "object"
      },
...
      "ExpirySmile": {
        "properties": {
          "expiry": {
            "title": "Expiry",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/SmilePoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "expiry",
          "points"
        ],
        "title": "ExpirySmile",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionQuoteOut": {
        "properties": {
          "ask": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ask"
          },
          "bid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "expiry": {
            "title": "Expiry",
            "type": "string"
          },
          "iv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iv"
          },
          "multiplier": {
            "title": "Multiplier",
            "type": "number"
          },
          "right": {
            "title": "Right",
            "type": "string"
          },
          "strike": {
            "title": "Strike",
            "type": "number"
          }
        },
        "required": [
          "expiry",
          "strike",
          "right",
          "bid",
          "ask",
          "iv",
          "delta",
          "multiplier"
        ],
        "title": "OptionQuoteOut",
        "type": "object"
      },
...
      "PositionIn": {
        "description": "A book leg: qty != 0 shares/contracts of `symbol`. See module\ndocstring for the option-leg fields' optionality/defaulting convention.",
        "properties": {
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "expiry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry"
          },
          "multiplier": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multiplier"
          },
          "qty": {
            "title": "Qty",
            "type": "number"
          },
          "right": {
            "anyOf": [
              {
                "enum": [
                  "C",
                  "P"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Right"
          },
          "strike": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strike"
          },
          "symbol": {
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "qty"
        ],
        "title": "PositionIn",
        "type": "object"
      },
...
      "SmilePoint": {
        "properties": {
          "iv": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iv"
          },
          "strike": {
            "title": "Strike",
            "type": "number"
          }
        },
        "required": [
          "strike",
          "iv"
        ],
        "title": "SmilePoint",
        "type": "object"
      },
      "StressGridOut": {
        "properties": {
          "pnl": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Pnl",
            "type": "array"
          },
          "spot_shocks": {
            "items": {
              "type": "number"
            },
            "title": "Spot Shocks",
            "type": "array"
          },
          "vol_shocks": {
            "items": {
              "type": "number"
            },
            "title": "Vol Shocks",
            "type": "array"
          }
        },
        "required": [
          "vol_shocks",
          "spot_shocks",
          "pnl"
        ],
        "title": "StressGridOut",
        "type": "object"
      },
...
      "UnderlyingGreeksOut": {
        "properties": {
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "dollar_delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dollar Delta"
          },
          "gamma": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gamma"
          },
          "spot": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot"
          },
          "spy_equivalent_notional": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spy Equivalent Notional"
          },
          "theta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theta"
          },
          "underlier": {
            "title": "Underlier",
            "type": "string"
          },
          "vega": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vega"
          }
        },
        "required": [
          "underlier",
          "spot",
          "delta",
          "gamma",
          "vega",
          "theta",
          "dollar_delta",
          "spy_equivalent_notional"
        ],
        "title": "UnderlyingGreeksOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QuantMind API",
    "version": "0.4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/options/book-greeks": {
      "post": {
        "operationId": "book_greeks_api_options_book_greeks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookGreeksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookGreeksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Book Greeks"
      }
    },
    "/api/options/{underlier}/chain": {
      "get": {
        "operationId": "get_chain_api_options__underlier__chain_get",
        "parameters": [
          {
            "in": "path",
            "name": "underlier",
            "required": true,
            "schema": {
              "title": "Underlier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chain"
      }
    },
...
  }
}